Move contacts fetch logic
